# Conflicts:
#	api/src/main/java/venue/hub/api/domain/services/LoginService.java
#	api/src/main/java/venue/hub/api/domain/services/ProposalService.java
#	api/src/main/java/venue/hub/api/domain/services/VenueService.java

+++ api/src/main/java/venue/hub/api/domain/services/VenueService.java
@@ -4,7 +4,9 @@ import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import venue.hub.api.domain.dtos.event.EventResponseDTO;
 import venue.hub.api.domain.dtos.mapper.EventMapper;
@@ -12,11 +14,14 @@ import venue.hub.api.domain.dtos.mapper.VenueMapper;
 import venue.hub.api.domain.dtos.venue.VenueRequestDTO;
 import venue.hub.api.domain.dtos.venue.VenueResponseDTO;
 import venue.hub.api.domain.dtos.venue.VenueUpdateDTO;
+import venue.hub.api.domain.dtos.venueadditional.VenueAdditionalRemoveDTO;
 import venue.hub.api.domain.dtos.venueadditional.VenueAdditionalRequestDTO;
 import venue.hub.api.domain.entities.*;
+import venue.hub.api.domain.enums.UserRole;
 import venue.hub.api.domain.repositories.AddressRepository;
 import venue.hub.api.domain.repositories.VenueAdditionalRepository;
 import venue.hub.api.domain.repositories.VenueRepository;
+import venue.hub.api.domain.specification.VenueSpecification;
 import venue.hub.api.domain.validators.address.AddressValidator;
 import venue.hub.api.infra.exceptions.VenueAdditionalNotFound;
 import venue.hub.api.infra.exceptions.VenueNotFoundException;
@@ -58,6 +63,9 @@ public class VenueService {
 
         Venue venue = venueMapper.toEntity(requestDTO);
 
+        User user = authenticationService.getAuthenticatedUser();
+        venue.setUser(user);
+
         addressRepository.save(venue.getAddress());
         venueRepository.save(venue);
 
@@ -72,42 +80,47 @@ public class VenueService {
             }
             venue.setAdditionals(additionalList);
         }
+
         return venueMapper.toDTO(venue);
     }
 
-    public Page<VenueResponseDTO> getVenuesByUser(Pageable paginacao) {
+    public Page<VenueResponseDTO> getAllVenues(Specification<Venue> spec, Pageable paginacao) {
         User user = authenticationService.getAuthenticatedUser();
 
         switch (user.getRole()) {
-            case OWNER -> {
-                return venueRepository.findByUser(user, paginacao)
-                        .map(venueMapper::toDTO);
-            }
-            case ADMIN -> {
-                return getAllVenues(paginacao);
-            }
+            case OWNER -> spec = spec.and(VenueSpecification.comOwner(user));
+            case ADMIN -> spec = spec.and(VenueSpecification.comOwner(null));
             default -> {
                 return null;
             }
         }
 
-    }
-
-    private Page<VenueResponseDTO> getAllVenues(Pageable paginacao) {
-        return venueRepository.findAllByAtivoTrue(paginacao)
+        return venueRepository.findAll(spec, paginacao)
                 .map(venueMapper::toDTO);
     }
 
     public VenueResponseDTO getVenueById(Long id) {
         var venue = this.findById(id);
+        var user = authenticationService.getAuthenticatedUser();
+        if (!venue.getUser().equals(user) && user.getRole() != UserRole.ADMIN) {
+            throw new AccessDeniedException("Usuário autenticado não é dono da Venue (" + venue.getUser().getLogin() + ")");
+        }
 
         return venueMapper.toDTO(venue);
     }
 
+    @Transactional
     public VenueResponseDTO updateVenue(Long id, VenueUpdateDTO updateDTO) {
         var venue = this.findById(id);
+        var user = authenticationService.getAuthenticatedUser();
+
+        if (!venue.getUser().equals(user) && user.getRole() != UserRole.ADMIN) {
+            throw new AccessDeniedException("Usuário autenticado não é dono da Venue (" + venue.getUser().getLogin() + ")");
+        }
+
         venue.update(updateDTO);
 
+        venueRepository.save(venue);
         addressRepository.save(venue.getAddress());
 
         return venueMapper.toDTO(venue);
@@ -115,20 +128,23 @@ public class VenueService {
 
     public void deleteVenue(Long id) {
         Venue venue = this.findById(id);
-        venue.setAtivo(false);
+        var user = authenticationService.getAuthenticatedUser();
+
+        if (!venue.getUser().equals(user) && user.getRole() != UserRole.ADMIN) {
+            throw new AccessDeniedException("Usuário autenticado não é dono da Venue (" + venue.getUser().getLogin() + ")");
+        }
+
+        venue.delete();
         venueRepository.save(venue);
     }
 
     public Page<EventResponseDTO> getEvents(Long venueId, Pageable paginacao) {
-        User user = authenticationService.getAuthenticatedUser();
-
-        return venueRepository.findEvents(venueId, user, paginacao)
+        return venueRepository.findEvents(venueId, paginacao)
                 .map(eventMapper::toDTO);
     }
 
     public Page<EventResponseDTO> getEventsByMonthAndYear(Long venueId, int month, int year, Pageable paginacao) {
-        User user = authenticationService.getAuthenticatedUser();
-        return venueRepository.findConfirmedEventsByVenueAndDateAndUser(venueId, month, year, user, paginacao)
+        return venueRepository.findConfirmedEventsByVenueAndDate(venueId, month, year, paginacao)
                 .map(eventMapper::toDTO);
     }
 
@@ -137,8 +153,13 @@ public class VenueService {
                 .orElseThrow(() -> new VenueNotFoundException(HttpStatus.NOT_FOUND, "Local não encontrado com o id: " + id));
     }
 
-    public VenueResponseDTO updateVenueAdditionals(Long venueId, List<VenueAdditionalRequestDTO> additionals) {
-        Venue venue = findById(venueId);
+    public VenueResponseDTO addAdditionalsToVenue(Long venueId, List<VenueAdditionalRequestDTO> additionals) {
+        Venue venue = this.findById(venueId);
+        var user = authenticationService.getAuthenticatedUser();
+
+        if (!venue.getUser().equals(user) && user.getRole() != UserRole.ADMIN) {
+            throw new AccessDeniedException("Usuário autenticado não é dono da Venue (" + venue.getUser().getLogin() + ")");
+        }
 
         List<VenueAdditional> venueAdditionals = new ArrayList<>();
         for (VenueAdditionalRequestDTO additionalDTO : additionals) {
@@ -148,7 +169,9 @@ public class VenueService {
 
             venueAdditionals.add(venueAdditional);
         }
+
         venueAdditionalRepository.saveAll(venueAdditionals);
+
         return venueMapper.toDTO(venue);
     }
 
@@ -161,17 +184,24 @@ public class VenueService {
         return venueAdditional;
     }
 
-    public void removeAdditionalFromVenue(Long venueId, List<Long> additionalIds) {
-        Venue venue = findById(venueId);
+    public VenueResponseDTO removeAdditionalsFromVenue(Long venueId, List<VenueAdditionalRemoveDTO> additionals) {
+        Venue venue = this.findById(venueId);
+        var user = authenticationService.getAuthenticatedUser();
+
+        if (!venue.getUser().equals(user) && user.getRole() != UserRole.ADMIN) {
+            throw new AccessDeniedException("Usuário autenticado não é dono da Venue (" + venue.getUser().getLogin() + ")");
+        }
 
         List<VenueAdditional> toRemove = new ArrayList<>();
-        for (Long additionalId : additionalIds) {
-            VenueAdditionalId venueAdditionalId = new VenueAdditionalId(venue.getId(), additionalId);
+        for (VenueAdditionalRemoveDTO additionalDTO : additionals) {
+            VenueAdditionalId venueAdditionalId = new VenueAdditionalId(venue.getId(), additionalDTO.getAdditionalId());
             VenueAdditional venueAdditional = venueAdditionalRepository.findById(venueAdditionalId)
-                    .orElseThrow(() -> new VenueAdditionalNotFound("Adicional não encontrado com o id: " + additionalId, HttpStatus.NOT_FOUND));
+                    .orElseThrow(() -> new VenueAdditionalNotFound("Adicional não encontrado com o id: " + additionalDTO.getAdditionalId(), HttpStatus.NOT_FOUND));
             toRemove.add(venueAdditional);
         }
         venueAdditionalRepository.deleteAll(toRemove);
         venue.getAdditionals().removeAll(toRemove);
+
+        return venueMapper.toDTO(venue);
     }
 }
